Move StringEncoder and StringDecoder into the wasm package
They aren't needed on jvm.

Made a special testonly package containing the required jvm deps needed
to be able to unit test those wasm classes
