+++ java/arcs/sdk/wasm/WasmCollection.kt
@@ -12,7 +12,6 @@
 package arcs.sdk.wasm
 
 import arcs.sdk.ReadWriteCollection
-import arcs.sdk.StringDecoder
 
 /** [ReadWriteCollection] implementation for WASM. */
 class WasmCollection<T : WasmEntity>(
