+++ javatests/arcs/sdk/wasm/BUILD
@@ -1,11 +1,27 @@
 load(
     "//third_party/java/arcs/build_defs:build_defs.bzl",
+    "arcs_kt_jvm_test_suite",
     "arcs_kt_library",
     "arcs_kt_particles",
     "arcs_kt_schema",
 )
 
 licenses(["notice"])
+
+WASM_UNIT_TEST_SRCS = [
+    "StringDecoderTest.kt",
+    "StringEncoderTest.kt",
+]
+
+PARTICLE_TEST_BASE_SRCS = [
+    "Test.kt",
+    "TestBase.kt",
+]
+
+PARTICLE_TEST_SRCS = glob(
+    ["*.kt"],
+    exclude = PARTICLE_TEST_BASE_SRCS + WASM_UNIT_TEST_SRCS,
+)
 
 arcs_kt_schema(
     name = "schemas",
@@ -15,13 +31,7 @@ arcs_kt_schema(
 
 arcs_kt_particles(
     name = "test-module",
-    srcs = glob(
-        ["*.kt"],
-        exclude = [
-            "TestBase.kt",
-            "Test.kt",
-        ],
-    ),
+    srcs = PARTICLE_TEST_SRCS,
     jvm = False,
     package = "arcs.sdk.wasm",
     visibility = ["//visibility:public"],
@@ -33,13 +43,21 @@ arcs_kt_particles(
 
 arcs_kt_library(
     name = "TestBase",
-    srcs = [
-        "Test.kt",
-        "TestBase.kt",
-    ],
+    srcs = PARTICLE_TEST_BASE_SRCS,
     jvm = False,
     deps = [
         ":schemas",
         "//java/arcs/sdk:arcs",
+    ],
+)
+
+arcs_kt_jvm_test_suite(
+    name = "WasmUnitTests",
+    srcs = WASM_UNIT_TEST_SRCS,
+    package = "arcs.sdk.wasm",
+    deps = [
+        "//java/arcs/sdk:arcs-wasm-testing",
+        "//third_party/java/junit:junit-android",
+        "//third_party/java/truth:truth-android",
     ],
 )
