+++ javatests/arcs/sdk/wasm/StringEncoderTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Google LLC.
+ * Copyright 2020 Google LLC.
  *
  * This code may only be used under the BSD style license found at
  * http://polymer.github.io/LICENSE.txt
@@ -9,7 +9,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-package arcs.sdk
+package arcs.sdk.wasm
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
