+++ src/tools/tests/goldens/generated-schemas.kt
@@ -17,12 +17,12 @@ import arcs.sdk.StringDecoder
 import arcs.sdk.StringEncoder
 import arcs.sdk.WritableCollection
 import arcs.sdk.WritableSingleton
-import arcs.sdk.toUtf8String
 import arcs.sdk.wasm.WasmCollection
 import arcs.sdk.wasm.WasmEntity
 import arcs.sdk.wasm.WasmEntitySpec
 import arcs.sdk.wasm.WasmParticle
 import arcs.sdk.wasm.WasmSingleton
+import arcs.sdk.wasm.toUtf8String
 
 class GoldInternal1() : WasmEntity {
